Parte lucas + main, no compila aun, handlers separados
