+++ .vscode/settings.json
@@ -51,6 +51,7 @@
         "text_encoding": "cpp",
         "typeinfo": "cpp",
         "variant": "cpp",
-        "map": "cpp"
+        "map": "cpp",
+        "cstdarg": "cpp"
     }
 }
